+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.2.0] - 2025-12-12
+### Fixed
+- Use generic H2 concentration encoding/decoding functions
+- Fix clang-tidy warnings
+
 ## [1.1.1] - 2025-11-26
 ### Fixed
 - Corrected H2 concentration encoding/decoding functions